pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiosqlite = "^0.20.0"
PyJWT = "^2.8.0"
fastapi = {extras = ["all"], version = "^0.110.3"}
sqlalchemy = "^2.0.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
